Update stuff to .NET 6 (#7688)
* Update to .NET 6

* Update templates

* Re-enable building of mgcb editor for Mac

* Remove the nuget.config that was used for XamMac2 as we don't need it as we use MonoMac

* Ups, forgot about some templates

* fix

* Remove some deprecated calls from build.cake

* Fix up the MGCB.Task and update the templates to install the dotnet mgcb tool

* Use env variable for dotnet

* template fixes

* remove new line so that the file does not count as modified on changes

* Add LangVersion 5 to WindowsDX

+++ MonoGame.Framework/Platform/Input/GamePad.iOS.cs
@@ -17,7 +17,7 @@ namespace Microsoft.Xna.Framework.Input
         static bool IndexIsUsed(GCControllerPlayerIndex index)
         {
             foreach (var ctrl in GCController.Controllers)
-                if (ctrl.PlayerIndex == (int)index) return true;
+                if (ctrl.PlayerIndex == index) return true;
 
             return false;
         }
@@ -28,11 +28,11 @@ namespace Microsoft.Xna.Framework.Input
                 return;
             foreach (var controller in GCController.Controllers)
             {
-                if (controller.PlayerIndex == (int)index)
+                if (controller.PlayerIndex == index)
                     break;
-                if (controller.PlayerIndex == (int)GCControllerPlayerIndex.Unset)
+                if (controller.PlayerIndex == GCControllerPlayerIndex.Unset)
                 {
-                    controller.PlayerIndex = (int)index;
+                    controller.PlayerIndex = index;
                     break;
                 }
             }
@@ -48,7 +48,7 @@ namespace Microsoft.Xna.Framework.Input
             {
                 if (controller == null)
                     continue;
-                if (controller.PlayerIndex == (int)ind)
+                if (controller.PlayerIndex == ind)
                     return GetCapabilities(controller);
             }
             return new GamePadCapabilities { IsConnected = false };
@@ -125,7 +125,7 @@ namespace Microsoft.Xna.Framework.Input
                 if (controller == null)
                     continue;
 
-                if (controller.PlayerIndex != (int)ind)
+                if (controller.PlayerIndex != ind)
                     continue;
 
                 connected = true;
